# app/openapi.yaml
openapi: 3.0.0
info:
  title: Petnet API's Doc
  description: This file content all the routes on Petnet application.
  version: 1.0.0
servers:
  - url: "https://petnet-8e037ce96247.herokuapp.com"
    description: Main production server.
  - url: "htpp://localhost:5000"
    description: Internal staging server for development.
...
paths:
...
  /users/information:
    get:
      tags:
        - Petnet-Users
      summary: Get user information.
      description: Retrieves user information.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: User information retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    description: The ID of the user.
                  first_name:
                    type: string
                    description: The first_name of the user.
                  last_name:
                    type: string
                    description: The last_name of the user.
                  email:
                    type: string
                    description: The email of the user.
                  clinic:
                    type: string
                    description: The clinic name of the user
                  username:
                    type: string
                    description: The username of the user.
                  confirmed:
                    type: boolean
                    description: Bool if the user is confirmed.
        '400':
          description: Bad Request. Error retrieving user information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure to retrieve user information.
                  error:
                    type: string
                    description: Additional information about the error.
        '401':
          description: Unauthorized. Invalid or missing token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to invalid or missing token.
  /users/profile_photo/{user_id}:
    get:
      tags:
        - Petnet-Users
      summary: Get user profile photo
      description: Retrieves and serves the profile photo of a user.
      parameters:
        - in: path
          name: user_id
          required: true
          description: The ID of the user.
          schema:
            type: integer
      responses:
        '200':
          description: Profile photo served successfully.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
                description: The profile photo of the user.
        '404':
          description: Not Found. User or profile photo not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to user not found or no profile photo available.
  /users/update/information:
    put:
      tags:
        - Petnet-Users
      summary: Update user information.
      description: Updates user information.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: The first name of the user.
                last_name:
                  type: string
                  description: The last name of the user.
                clinic:
                  type: string
                  description: The clinic name associated with the user.
                username:
                  type: string
                  description: The username chosen by the user.
                address:
                  type: string
                  description: The address of the clinic.
                college_number:
                  type: string
                  description: Collegue number of the vet doc.
                
              required:
                - first_name
                - last_name
                - clinic
                - address
                - college_number
      responses:
        '200':
          description: User information updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful user information update.
        '400':
          description: Bad Request. Error updating user information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure to update user information.
                  error:
                    type: string
                    description: Additional information about the error.
        '401':
          description: Unauthorized. Invalid or missing token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to invalid or missing token.
  /users/update/password:
    put:
      tags:
        - Petnet-Users
      summary: Update user password
      description: Updates the password of the user.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actual_password:
                  type: string
                  format: password
                  description: The current password of the user.
                new_password:
                  type: string
                  format: password
                  description: The new password for the user.
                confirm_password:
                  type: string
                  format: password
                  description: The confirmation of the new password.
              required:
                - actual_password
                - new_password
                - confirm_password
      responses:
        '200':
          description: User password updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful password update.
        '400':
          description: Bad Request. Error updating user password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure to update user password.
                  error:
                    type: string
                    description: Additional information about the error.
        '401':
          description: Unauthorized. Invalid or missing token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to invalid or missing token.   
  /users/restore/password/verify:
    post:
      tags:
        - Petnet-Users
      summary: Verify restored password
      description: Restore the password using the code sent to the user email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: The verification code sent to the user email.
                token:
                  type: string
                  description: The token associated with the code verification.
              required:
                - code
                - token
      responses:
        '200':
          description: Code verified successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful code verification.
                  user_id:
                    type: string
                    description: The ID of the user for whom the code was verified.
        '400':
          description: Bad Request. Invalid code or token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to invalid code or token.
        '404':
          description: Not Found. User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to user not found.
  /users/restore/password/update:
    post:
      tags:
        - Petnet-Users
      summary: Update restored password
      description: Update the password using the code sent to the user email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: The ID of the user.
                new_password:
                  type: string
                  format: password
                  description: The new password for the user.
                confirm_password:
                  type: string
                  format: password
                  description: The confirmation of the new password.
              required:
                - user_id
                - new_password
                - confirm_password
      responses:
        '200':
          description: Password updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful password update.
        '400':
          description: Bad Request. Error updating user password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure to update user password.
                  error:
                    type: string
                    description: Additional information about the error.
        '404':
          description: Not Found. User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to user not found.
  /users/update/email:
    post:
      tags:
        - Petnet-Users
      summary: Update user email.
      description: Update the email address of the user.

      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_email:
                  type: string
                  format: email
                  description: The new email address for the user.
              required:
                - new_email
      responses:
        '200':
          description: Email update requested successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful email update request.
        '400':
          description: Bad Request. Error updating user email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure to update user email.
                  error:
                    type: string
                    description: Additional information about the error.
        '404':
          description: Not Found. User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to user not found.
  /users/delete:
    delete:
      tags:
        - Petnet-Users
      summary: Delete user.
      description: Deletes a user.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: User deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful user deletion.
        '500':
          description: Internal Server Error. Error deleting user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure to delete user.
                  error:
                    type: string
                    description: Additional information about the error.
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization 
security:
  - ApiKeyAuth: []